Add uav::optional<T>
No, this isn't std::optional<T>

The difference is uav::optional<T> always holds a value.
The boolean flag can indicate whatever you want, and is intended
to be context-dependent. The idea for this is to use it to
show if a returned value is crap or not, like in the event of
bad sensor readings.